+++ src/array/array_metadata_options.rs
@@ -4,6 +4,16 @@ use crate::config::global_config;
 #[derive(Debug, Clone)]
 pub struct ArrayMetadataOptions {
     experimental_codec_store_metadata_if_encode_only: bool,
+    version: ArrayMetadataOptionsVersion,
+}
+
+/// Array [`Array::store_metadata`](crate::array::Array::store_metadata) version options.
+#[derive(Debug, Clone, Copy)]
+pub enum ArrayMetadataOptionsVersion {
+    /// Write the same version as the input metadata.
+    Unchanged,
+    /// Write v3. Replace v2 if it exists.
+    V3,
 }
 
 impl Default for ArrayMetadataOptions {
@@ -11,6 +21,7 @@ impl Default for ArrayMetadataOptions {
         Self {
             experimental_codec_store_metadata_if_encode_only: global_config()
                 .experimental_codec_store_metadata_if_encode_only(),
+            version: *global_config().array_metadata_version(),
         }
     }
 }
@@ -25,5 +36,16 @@ impl ArrayMetadataOptions {
     /// Set the [experimental codec store metadata if encode only](crate::config::Config#experimental-codec-store-metadata-if-encode-only) setting.
     pub fn set_experimental_codec_store_metadata_if_encode_only(&mut self, enabled: bool) {
         self.experimental_codec_store_metadata_if_encode_only = enabled;
+    }
+
+    /// Get the [array metadata version behaviour](crate::config::Config#array-metadata-version-behaviour) configuration.
+    #[must_use]
+    pub fn array_metadata_version(&self) -> &ArrayMetadataOptionsVersion {
+        &self.version
+    }
+
+    /// Set the [array metadata version behaviour](crate::config::Config#array-metadata-version-behaviour) configuration.
+    pub fn set_array_metadata_version(&mut self, version: ArrayMetadataOptionsVersion) {
+        self.version = version;
     }
 }